+++ exporters/trace/jaeger/uploader.go
@@ -17,7 +17,6 @@ package jaeger // import "go.opentelemetry.io/otel/exporters/trace/jaeger"
 import (
 	"bytes"
 	"context"
-	"errors"
 	"fmt"
 	"io"
 	"io/ioutil"
@@ -114,30 +113,30 @@ func WithAttemptReconnectingInterval(interval time.Duration) AgentEndpointOption
 	}
 }
 
-// WithCollectorEndpoint defines the full url to the Jaeger HTTP Thrift collector.
-// For example, http://localhost:14268/api/traces
-func WithCollectorEndpoint(collectorEndpoint string, options ...CollectorEndpointOption) EndpointOption {
+// WithCollectorEndpoint defines the full url to the Jaeger HTTP Thrift collector. This will
+// use the following environment variables for configuration if no explicit option is provided:
+//
+// - OTEL_EXPORTER_JAEGER_ENDPOINT is the HTTP endpoint for sending spans directly to a collector
+// - OTEL_EXPORTER_JAEGER_USER is the username to be sent as "Basic" authentication to the collector endpoint.
+// - OTEL_EXPORTER_JAEGER_PASSWORD is the password to be sent as "Basic" authentication to the collector endpoint.
+//
+// The passed options will take precedence over any environment variables. Default values will be
+// used if neither are provided.
+func WithCollectorEndpoint(options ...CollectorEndpointOption) EndpointOption {
 	return func() (batchUploader, error) {
-		// Overwrite collector endpoint if environment variables are available.
-		if e := CollectorEndpointFromEnv(); e != "" {
-			collectorEndpoint = e
-		}
-
-		if collectorEndpoint == "" {
-			return nil, errors.New("collectorEndpoint must not be empty")
-		}
-
 		o := &CollectorEndpointOptions{
+			endpoint:   envOr(envEndpoint, "http://localhost:14250"),
+			username:   envOr(envUser, ""),
+			password:   envOr(envPassword, ""),
 			httpClient: http.DefaultClient,
 		}
 
-		options = append(options, WithCollectorEndpointOptionFromEnv())
 		for _, opt := range options {
 			opt(o)
 		}
 
 		return &collectorUploader{
-			endpoint:   collectorEndpoint,
+			endpoint:   o.endpoint,
 			username:   o.username,
 			password:   o.password,
 			httpClient: o.httpClient,
@@ -148,6 +147,9 @@ func WithCollectorEndpoint(collectorEndpoint string, options ...CollectorEndpoin
 type CollectorEndpointOption func(o *CollectorEndpointOptions)
 
 type CollectorEndpointOptions struct {
+	// endpoint for sending spans directly to a collector
+	endpoint string
+
 	// username to be used if basic auth is required.
 	username string
 
@@ -158,7 +160,20 @@ type CollectorEndpointOptions struct {
 	httpClient *http.Client
 }
 
+// WithEndpoint is the HTTP endpoint for sending spans directly to a collector
+// This option overrides any value set for the
+// OTEL_EXPORTER_JAEGER_ENDPOINT environment variable.
+// If this option is not passed and the env var is not set, "http://localhost:14250" will be used by default.
+func WithEndpoint(endpoint string) CollectorEndpointOption {
+	return func(o *CollectorEndpointOptions) {
+		o.endpoint = endpoint
+	}
+}
+
 // WithUsername sets the username to be used if basic auth is required.
+// This option overrides any value set for the
+// OTEL_EXPORTER_JAEGER_USER environment variable.
+// If this option is not passed and the env var is not set, "" will be used by default.
 func WithUsername(username string) CollectorEndpointOption {
 	return func(o *CollectorEndpointOptions) {
 		o.username = username
@@ -166,6 +181,9 @@ func WithUsername(username string) CollectorEndpointOption {
 }
 
 // WithPassword sets the password to be used if basic auth is required.
+// This option overrides any value set for the
+// OTEL_EXPORTER_JAEGER_PASSWORD environment variable.
+// If this option is not passed and the env var is not set, "" will be used by default.
 func WithPassword(password string) CollectorEndpointOption {
 	return func(o *CollectorEndpointOptions) {
 		o.password = password